% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tokens.R
\name{count_tokens}
\alias{count_tokens}
\title{counts the frequency of unique tokens}
\usage{
count_tokens(x)
}
\arguments{
\item{x}{length 1 character vector}
}
\value{
A table of word frequencies
}
\description{
Call word tokenizer, then count the frequency of each unique word.
}
\examples{
count_tokens("If this function is working right, it is going to return a count of three for the word, 'is.'")
}
\seealso{
\code{\link{str_split}} \code{\link{boundary}} \code{\link{tokenize}}
}
