% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment}
\alias{sentiment}
\title{Conduct dictionary-based sentiment analysis.}
\usage{
sentiment(x,tokens,scores)
}
\arguments{
\item{x}{length n character vector. If you're input dictionary is lower case only, then be sure to preprocess inputs to x using tolower.}

\item{tokens}{length m character vector containing dictionary words}

\item{scores}{m x d numerical matrix containing scores for each dictionary word on d-dimensions. Values can be weighted continuous scores or binary.}
}
\value{
n x d numerical matrix with scores on x for dimensions d
}
\description{
Tokenizes character vectors, counts words, and then takes weighted average to estimate sentiment.
}
\examples{
dictionary <- c("great","good","bad","awful")
scores <- matrix(c(1,.5,-.5,-1),4,1)
sentiment("This package is great.",dictionary,scores)
sentiment("This package is good.",dictionary,scores)
sentiment("This package is bad.",dictionary,scores)
sentiment("This package is awful.",dictionary,scores)
}
\seealso{
\code{\link{count_tokens}} \code{\link{tokenize}}
}
